add test if default cache is using django-redis.

+++ course/receivers.py
@@ -109,6 +109,7 @@ def may_preapprove_role(course, user):
 
 # }}}
 
+
 # {{{ Flush redis cache on event save or delete
 
 @receiver(post_delete, sender=Event)
@@ -136,12 +137,15 @@ def flush_event_cache(course):
     except ImproperlyConfigured:
         return
 
+    def_cache = cache.caches["default"]
+    if not hasattr(def_cache, "delete_pattern"):
+        return
+
     from course.constants import DATESPECT_CACHE_KEY_PATTERN
     cache_pattern = DATESPECT_CACHE_KEY_PATTERN % {
         "course": course.identifier,
         "key": "*"
     }
-    def_cache = cache.caches["default"]
     try:
         def_cache.delete_pattern(cache_pattern)
     except Exception as e:
